TestReflect running, not yet passing

+++ src/std/buffer.c
@@ -64,7 +64,7 @@ void hl_buffer_str( hl_buffer *b, const uchar *s ) {
 
 void hl_buffer_cstr( hl_buffer *b, const char *s ) {
 	if( s ) {
-		int len = strlen(s);
+		int len = (int)strlen(s);
 		uchar *out = (uchar*)malloc(sizeof(uchar)*(len+1));
 		strtou(out,len,s);
 		hl_buffer_str_sub(b,out,len);
